feat: add nvidia kmod builds and ublue nvidia support RPM (#3)
Builds both nvidia driver versions 470 and 535 with a 'latest' symlink
to version 535 allowing downstream usages to use a non-version specific
path.

Note: dnf is used in the nvidia builds to allow uninstalls to prevent
version collisions.

Also builds an RPM with ublue-os/ucore specific extra files to support
nvidia container runtime, selinux policy, and rootless containers. This
is a clone of the ublue-os-nvidia-addons RPM from ublue-os/nvidia but
stripped down for the Fedora CoreOS use case.